Refactor to extract common code from extensions.
Uses ActiveSupport::Concern to control the addition of methods.
The Kaminari public API (scope DSL, class methods for config)
was extracted to PageScopeMethods and ConfigurationMethods.
Shims to abstract the difference between Mongoid and ActiveRecord
were extracted to MongoidCriteriaMethods and ActiveRecordRelationMethods
respectively.

Also moved definition of common constants up to the railtie.